updated admin overview page to seperate components to use bot admin and agent page
